ft(delete question): added delete question functionalit by - added delete functionality in context API - delete functionality on view

+++ src/context/AuthContext.tsx
@@ -1,8 +1,9 @@
 /** @format */
 
 import { createContext, useEffect, useState } from "react";
-import { AuthContextState, IQuestion} from "./Types";
+import { AuthContextState, IQuestion } from "./Types";
 import { auth } from "../firebase";
+import Swal from "sweetalert2";
 
 const contextDefaultValue: AuthContextState = {
 	currentUser: "",
@@ -112,11 +113,22 @@ const AuthProvider = ({ children }: AuthContextProviderProps) => {
 
 	// Delete Question
 	const deleteQuestion = (questionId: number) => {
-		fetch(`http://localhost:8000/questions/` + questionId, {
-			method: "DELETE",
-		}).then(() => {
-			//action done
-			getAllQuestions();
+		Swal.fire({
+			title: "Do you want to delete question",
+			showDenyButton: true,
+			confirmButtonText: "Yes",
+			denyButtonText: `No`,
+		}).then((result) => {
+			/* Read more about isConfirmed, isDenied below */
+			if (result.isConfirmed) {
+				fetch(`http://localhost:8000/questions/` + questionId, {
+					method: "DELETE",
+				}).then(() => {
+					//action done
+					getAllQuestions();
+				});
+			} else if (result.isDenied) {
+			}
 		});
 	};
